requirements.txt
pandas
numpy
tqdm
jupyter
matplotlib
seaborn
openpyxl
click
python-dotenv
